server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.2
slowapi==0.1.9
alembic==1.13.2
python-docx==1.1.2
pypdf==5.0.1
reportlab==4.2.2
pytest==8.3.3
